news-parser/requirements.txt
requests
beautifulsoup4
pandas
PyYAML
pytest
